Create FileBasedDataSourceSuite.
