+++ sql/core/src/test/scala/org/apache/spark/sql/SQLQuerySuite.scala
@@ -2757,38 +2757,4 @@ class SQLQuerySuite extends QueryTest with SharedSQLContext {
       }
     }
   }
-
-  // Only New OrcFileFormat supports this
-  Seq(classOf[org.apache.spark.sql.execution.datasources.orc.OrcFileFormat].getCanonicalName,
-      "parquet").foreach { format =>
-    test(s"SPARK-15474 Write and read back non-emtpy schema with empty dataframe - $format") {
-      withTempPath { file =>
-        val path = file.getCanonicalPath
-        val emptyDf = Seq((true, 1, "str")).toDF.limit(0)
-        emptyDf.write.format(format).save(path)
-
-        val df = spark.read.format(format).load(path)
-        assert(df.schema.sameType(emptyDf.schema))
-        checkAnswer(df, emptyDf)
-      }
-    }
-  }
-
-  Seq("orc", "parquet", "csv", "json").foreach { format =>
-    test(s"Write and read back unicode schema - $format") {
-      withTempPath { path =>
-        val dir = path.getCanonicalPath
-
-        // scalastyle:off nonascii
-        val df = Seq("a").toDF("한글")
-        // scalastyle:on nonascii
-
-        df.write.format(format).option("header", "true").save(dir)
-        val answerDf = spark.read.format(format).option("header", "true").load(dir)
-
-        assert(df.schema === answerDf.schema)
-        checkAnswer(df, answerDf)
-      }
-    }
-  }
 }
